Revert "feat(pb): refactor PocketBase usage to composable (closes #12) (#13)"
This reverts commit 5749220a701347c807148826cc91867fcbf2346a.
