+++ web/components/AppNav.vue
@@ -1,14 +1,13 @@
 <script setup lang="ts">
 import type { NavigationMenuItem } from '@nuxt/ui'
-import { navigateTo, useNuxtApp } from '#app'
+import { navigateTo } from '#app'
 import {
+  pb,
   ref,
 } from '#imports'
 
-const { $pb } = useNuxtApp()
-
 const onLogout = async () => {
-  $pb.client.authStore.clear()
+  pb.authStore.clear()
   await navigateTo({ name: 'login' })
 }
 
@@ -21,8 +20,8 @@ const items = ref<NavigationMenuItem[]>([
   ],
   [
     {
-      label: $pb.client.authStore.isValid ? 'Logout' : 'Login',
-      onClick: $pb.client.authStore.isValid
+      label: pb.authStore.isValid ? 'Logout' : 'Login',
+      onClick: pb.authStore.isValid
         ? onLogout
         : () => navigateTo({ name: 'login' }),
     },
