+++ web/pages/login.vue
@@ -1,14 +1,12 @@
 <script lang="ts" setup>
 import { onMounted, reactive } from 'vue'
-import { navigateTo, useNuxtApp } from '#app'
-import { useToast } from '#imports'
+import { navigateTo } from '#app'
+import { useToast, pb } from '#imports'
 
 const state = reactive({
   email: '',
   password: '',
 })
-
-const { $pb } = useNuxtApp()
 
 const isAuthorized = defineModel('isAuthorized', {
   type: Boolean,
@@ -17,7 +15,7 @@ const isAuthorized = defineModel('isAuthorized', {
 
 function setAuthorized(value: boolean) {
   if (!value) {
-    $pb.client.authStore.clear()
+    pb.authStore.clear()
   }
   isAuthorized.value = value
 }
@@ -35,11 +33,11 @@ const onSubmit = async () => {
   }
 
   try {
-    await $pb.client
+    await pb
       .collection('_superusers')
       .authWithPassword(state.email, state.password)
     setAuthorized(true)
-    if ($pb.client.authStore.isValid) {
+    if (pb.authStore.isValid) {
       await navigateTo({ name: 'index' })
     }
   }
@@ -53,7 +51,7 @@ const onSubmit = async () => {
 }
 
 onMounted(() => {
-  if ($pb.client.authStore.isValid) {
+  if (pb.authStore.isValid) {
     setAuthorized(true)
   }
 })
