+++ web/stores/activities.ts
@@ -1,24 +1,21 @@
 import { defineStore } from 'pinia'
 import {
+  pb,
   type Item,
   type ItemType,
   shallowRef,
   transformItem,
   useFiltersStore,
-  useNuxtApp,
 } from '#imports'
 
-
 export const useActivitiesStore = defineStore('activities', () => {
-  const { $pb } = useNuxtApp()
-
   const items = shallowRef<Item[]>([])
   const itemTypes = shallowRef<Set<ItemType>>(new Set())
   const item = shallowRef<Item | undefined>(undefined)
   const filtersStore = useFiltersStore()
 
   const load = async () => {
-    const resultList = await $pb.client.collection('files').getList(1, 20, {
+    const resultList = await pb.collection('files').getList(1, 20, {
       filter: filtersStore.buildQuery(),
     })
     items.value = resultList.items.map((item) => {
