+++ web/utils/pb.ts
@@ -1,8 +1,10 @@
-import type { RecordModel } from 'pocketbase'
-
+import PocketBase, { type RecordModel } from 'pocketbase'
+import { useRuntimeConfig } from '#app'
 import type { Item } from '#imports'
 
+const config = useRuntimeConfig()
 
+export const pb = new PocketBase(config.public.apiBase)
 
 // TODO: fix type assertion with validation library
 export function transformItem(item: RecordModel): Item {
